// backend/plane-b/src/repositories/interfaces/corridor-repository.interface.d.ts
export type CorridorInput = {
    corridorId: string;
    sourceCountry: string;
    destCountry: string;
    sourceCurrency: string;
    destCurrency: string;
};
export interface ICorridorRepository {
    insertIfMissing(input: CorridorInput): Promise<void>;
    upsertCorridor(input: CorridorInput): Promise<void>;
}